+++ docs/realm.js
@@ -120,7 +120,7 @@ class Realm {
     /**
      * Return a configuration for a default synced Realm. The server URL for the user will be used as base for
      * the URL for the synced Realm. If no user is supplied, the current user will be used.
-     * @param {Realm.Sync.User} - an optional sync user
+     * @param {Realm.User} - an optional sync user
      * @throws {Error} if zero or multiple users are logged in
      * @returns {Realm~Configuration} - a configuration matching a default synced Realm.
      * @since 2.3.0
@@ -384,7 +384,7 @@ class Realm {
  *   If omitted, the schema will be read from the existing Realm file.
  * @property {number} [schemaVersion] - **Required** (and must be incremented) after
  *   changing the `schema`.
- * @property {Realm.Sync~SyncConfiguration} [sync] - Sync configuration parameters.
+ * @property {Realm.App.Sync~SyncConfiguration} [sync] - Sync configuration parameters.
  */
 
 /**