+++ types/app.d.ts
@@ -329,6 +329,16 @@ declare namespace Realm {
         readonly id: string;
 
         /**
+         * The provider type for the user.
+         */
+        readonly providerType: string;
+
+        /**
+         * The id of the device.
+         */
+        readonly deviceId: string | null;
+
+        /**
          * The state of the user.
          */
         readonly state: UserState;