+++ CHANGELOG.md
@@ -4,7 +4,18 @@ NOTE: Support for syncing with realm.cloud.io and/or Realm Object Server has bee
 
 NOTE: This version bumps the Realm file format to version 20. It is not possible to downgrade to earlier versions. Non-sync Realms will be upgraded automatically. Synced Realms can only be automatically upgraded if created with Realm JavaScript v10.0.0-beta.1 and above.
 
+### Breaking changes
+* Renamed `Realm.Credentials.custom()` to `Realm.Credentials.jwt()`. ([RJS-689](https://jira.mongodb.org/browse/RJS-689))
+* Renamed `Realm.User.remoteMongoClient()` to `Realm.User.mongoClient()`. ([RJS-689](https://jira.mongodb.org/browse/RJS-689))
+* Removed `Remote` prefix from all `MongoDB` related classes.
+* Renamed `Realm.Sync` to `Realm.App.Sync` including all methods and properties. ([RJS-689](https://jira.mongodb.org/browse/RJS-689))
+
 ### Enhancements
+* Added property `Realm.User.deviceId`. ([RJS-689](https://jira.mongodb.org/browse/RJS-689))
+* Added property `Realm.User.providerType`. ([RJS-689](https://jira.mongodb.org/browse/RJS-689))
+* Added property `Realm.User.identities`. ([RJS-689](https://jira.mongodb.org/browse/RJS-689))
+* Added property `name` to `Realm.MongoDB`, `Realm.MongoDB.Database` and `Realm.MongoDB.Collection`. ([RJS-689](https://jira.mongodb.org/browse/RJS-689))
+* Added methods `Realm.App.Sync.getSyncSession()` and `Realm.App.Sync.getAllSyncSessions()`. ([RJS-689](https://jira.mongodb.org/browse/RJS-689))
 * Allow synchronization on the nil partition by specifying `null` as `partitionValue` in sync configuration (RJS-726).
 * Added support for creating multiple Realm apps. ([#3072](https://github.com/realm/realm-js/issues/3072))
 * Added method `Realm.App.getCachedApp()`.